+++ nomic/embed.py
@@ -16,6 +16,7 @@ from tqdm import tqdm
 
 from .dataset import AtlasClass
 from .settings import *
+from .utils import resize_pil
 
 try:
     from gpt4all import CancellationError, Embed4All
@@ -343,17 +344,6 @@ def image_api_request(
         return response.json()
     else:
         raise Exception((response.status_code, response.text))
-
-
-def resize_pil(img):
-    width, height = img.size
-    # if image is too large, downsample before sending over the wire
-    max_width = 512
-    max_height = 512
-    if width > max_width or height > max_height:
-        downsize_factor = max(width // max_width, height // max_height)
-        img = img.resize((width // downsize_factor, height // downsize_factor))
-    return img
 
 
 def _is_valid_url(url):
